styl(part/status): page show
